sum_interview/requirements.txt
Django==4.0.4
python-decouple==3.6
djangorestframework==3.13.1
